Fix the undefined left-shift operation.
I used undefined compiler behavior: left shift a uint8_t with 8+ bits.
This is fine in 32-bit, but very dangerous in 64-bit since uint8_t is
extended to int64_t, which is SIGNED.

Force cast to necessary bit width.

BUG=chrome-os-partner:8465
TEST=on ZGB.

Reviewed-on: https://gerrit.chromium.org/gerrit/17927
Commit-Ready: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>
Tested-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>

+++ mec1308.c
@@ -441,7 +441,8 @@ int mec1308_probe_spi_flash(const char *name)
 	tmp8 |= 1;
 	sio_write(sio_port, 0x30, tmp8);	/* activate logical device */
 
-	mbx_idx = sio_read(sio_port, 0x60) << 8 | sio_read(sio_port, 0x61);
+	mbx_idx = (unsigned int)sio_read(sio_port, 0x60) << 8 |
+	                        sio_read(sio_port, 0x61);
 	mbx_data = mbx_idx + 1;
 	msg_pdbg("%s: mbx_idx: 0x%04x, mbx_data: 0x%04x\n",
 	         __func__, mbx_idx, mbx_data);
